Made itemSlot UI and ItemSelectButton UI

+++ src/ItemSelectButton.java
@@ -3,12 +3,17 @@ import java.awt.*;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import javax.imageio.ImageIO;
 
 public class ItemSelectButton extends Sprite {
     private String itemName;
+    private String itemImgPath;
+    private String descImgPath;
+    private String type;
+    private static final Map<String, String> TEXTURE_CACHE = new HashMap<>();
     // 아이템 이미지 설정
     private static final Map<String, Object> ITEM_CONFIG = new LinkedHashMap<>();
     // 설명 이미지 설정
@@ -28,22 +33,44 @@ public class ItemSelectButton extends Sprite {
         DESC_CONFIG.put("arc", 20);
         DESC_CONFIG.put("borderColor", Color.BLACK);
     }
-    public ItemSelectButton(int x, int y, String name, String itemImgPath, String descImgPath) {
+    public ItemSelectButton(int x, int y, String type, String name, String itemImgPath, String descImgPath) {
+        GameManager.isGamePaused = true;
         this.itemName = name;
-        BufferedImage buttonSkin = drawButtonSkin(name, itemImgPath, descImgPath);
+        this.itemImgPath = itemImgPath;
+        this.descImgPath = descImgPath;
+        this.type = type;
 
+        // 위치 설정
+        this.setX(x);
+        this.setY(y);
+
+        // 코스튬 추가a
         try {
-            File tempFile = File.createTempFile("temp_btn_" + name, ".png");
-            ImageIO.write(buttonSkin, "png", tempFile);
-            tempFile.deleteOnExit();
-            this.addCostume("normal", tempFile.getAbsolutePath());
+            String finalPath;
+
+            if (TEXTURE_CACHE.containsKey(name)) {
+                finalPath = TEXTURE_CACHE.get(name);
+            }
+            else {
+                BufferedImage buttonSkin = drawButtonSkin(name, itemImgPath, descImgPath);
+                File tempFile = File.createTempFile("temp_btn_" + name, ".png");
+                ImageIO.write(buttonSkin, "png", tempFile);
+                tempFile.deleteOnExit();
+                finalPath = tempFile.getAbsolutePath();
+                TEXTURE_CACHE.put(name, finalPath);
+            }
+
+            this.addCostume(name, finalPath);
+            this.switchCostume(name);
         } catch (IOException e) {
             e.printStackTrace();
         }
-        this.setX(x);
-        this.setY(y);
+
+        GameManager.Instance.registerButton(this);
     }
 
+
+    // 버튼 그림
     private BufferedImage drawButtonSkin(String text, String path1, String path2) {
         int w = 200;
         int h = 300;
@@ -92,8 +119,8 @@ public class ItemSelectButton extends Sprite {
         g2.setFont(new Font("Malgun Gothic", Font.BOLD, 18));
         FontMetrics fm = g2.getFontMetrics();
         int textX = (w - fm.stringWidth(text)) / 2;
-        // 텍스트 위치도 아이템과 설명 사이 적절한 곳으로 조정하고 싶다면 변수화 가능
 
+        // 텍스트 위치도 아이템과 설명 사이 적절한 곳으로 조정하고 싶다면 변수화 가능
         g2.drawString(text, textX, 140);
 
         try {
@@ -107,6 +134,8 @@ public class ItemSelectButton extends Sprite {
 
     @Override
     public void whenClicked() {
-        this.say(itemName + " 선택됨!");
+        GameManager.isGamePaused = false;
+        GameManager.Instance.addItem(GameManager.WEAPON_SLOT,type,itemImgPath);
+        GameManager.Instance.hideOtherButtons();
     }
 }
